[clang][NFC] Create CodeGenShared component for shared code with ClangIR
