+++ clang/lib/CodeGenShared/CMakeLists.txt
@@ -1,0 +1,11 @@
+set(LLVM_LINK_COMPONENTS
+  Support
+  )
+
+add_clang_library(clangCodeGenShared
+  ItaniumRTTIBuilder.cpp
+
+  DEPENDS
+  LINK_LIBS
+  clangAST
+  )
